+++ app/src/main/res/layout/activity_about.xml
@@ -6,7 +6,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".features.about.AboutActivity">
+        tools:context=".activities.AboutActivity">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/about_toolbar"
