+++ app/src/main/res/layout/activity_settings.xml
@@ -6,7 +6,7 @@
         xmlns:tools="http://schemas.android.com/tools"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
-        tools:context=".features.settings.SettingsActivity">
+        tools:context=".activities.SettingsActivity">
 
         <androidx.appcompat.widget.Toolbar
             android:id="@+id/settings_toolbar"
